Autenticação das rotas usando JWT
- Separação das rotas
- Mudança do nome da model (Pontuacao >> ACC)

+++ src/database/migrations/1605140065523-create_tipo_de_acc.ts
@@ -20,7 +20,7 @@ export class createTipoDeAcc1605140065523 implements MigrationInterface {
         },
         {
           name: 'pontos_por_unidade',
-          type: 'decimal'
+          type: 'float'
         },
         {
           name: 'limite_de_pontos',
